aviasales_hakaton/check_my_baggage/translation.py

from modeltranslation.translator import translator, TranslationOptions

from check_my_baggage.models import AirlineCompany


class AirlineCompanyTranslationOptions(TranslationOptions):
    fields = ['name', 'description', 'baggage_allowance_link']  # logo?


translator.register(AirlineCompany, AirlineCompanyTranslationOptions)
